feature/API-015 into develop

+++ api/src/controllers/workspace.controller.js
@@ -1,20 +1,62 @@
-const WorkspaceService = require('../services').WorkspaceService;
+const {ServiceService, WorkspaceService} = require('../services');
+const UserController = require('./user.controller');
+const ServiceController = require('./service.controller');
 
 class WorkspaceController {
+    /**
+     * Add one service in one workspace from id
+     *
+     * @param workspaceId {number}
+     * @param serviceId {number}
+     *
+     * @returns {Promise<boolean|undefined>}
+     */
+    async addOneServiceInOneWorkspaceFromId(workspaceId, serviceId) {
+        const service = await ServiceController.findOneServiceFromId(serviceId);
+        if (service) {
+            const workspace = await WorkspaceService.findOne({id: workspaceId});
+            if (workspace) {
+                await workspace.addService(serviceId);
+                return true;
+            }
+        }
+    }
+
     /**
      * Create a new workspace
      *
      * @param name {string}
      * @param description {string}
+     * @param userToken {string}
      *
-     * @returns {Promise<WorkspaceDTO>}
+     * @returns {Promise<WorkspaceDTO|null>}
      */
-    async createWorkspace(name, description) {
+    async createWorkspace(name, description, userToken) {
+        const user = await UserController.findOneUserFromToken(userToken);
+        if (!user) {
+            return null;
+        }
         const workspace = await WorkspaceService.create({
             name: name,
-            description: description
+            description: description,
+            user_id: user.id
         });
         return WorkspaceService.mapToDTO(workspace);
+    }
+
+    /**
+     * Find all services of one workspace from id
+     *
+     * @param id {number}
+     *
+     * @returns {Promise<ServiceDTO[]|undefined>}
+     */
+    async findAllServicesOfOneWorkspaceFromId(id) {
+        const workspace = await WorkspaceService.findOne({id: id});
+        if (workspace) {
+            const services = await workspace.getServices();
+            return services.map(service => ServiceService.mapToDTO(service));
+        }
     }
 
     /**
@@ -40,15 +82,38 @@ class WorkspaceController {
     }
 
     /**
+     * Remove one workspace's service from id
+     *
+     * @param workspaceId {number}
+     * @param serviceId {number}
+     *
+     * @returns {Promise<boolean|undefined>}
+     */
+    async removeOneServiceOfOneWorkspaceFromId(workspaceId, serviceId) {
+        const service = await ServiceController.findOneServiceFromId(serviceId);
+        if (service) {
+            const workspace = await WorkspaceService.findOne({id: workspaceId});
+            if (workspace) {
+                await workspace.removeService(serviceId);
+                return true;
+            }
+        }
+    }
+
+    /**
      * Remove one workspace from id
      *
      * @param id {number}
+     * @param userToken {string}
      *
-     * @returns {Promise<boolean>}
+     * @returns {Promise<boolean|undefined>}
      */
-    async removeOneWorkspaceFromId(id) {
-        const workspace = await this.findOneWorkspaceFromId(id);
-        return !workspace ? false : await WorkspaceService.destroy({id: id});
+    async removeOneWorkspaceFromId(id, userToken) {
+        const user = await UserController.findOneUserFromToken(userToken);
+        if (user) {
+            const workspace = await this.findOneWorkspaceFromId(id);
+            return !workspace || workspace.user_id !== user.id ? false : await WorkspaceService.destroy({id: id});
+        }
     }
 
     /**
@@ -57,25 +122,29 @@ class WorkspaceController {
      * @param id {number}
      * @param name {string}
      * @param description {string}
+     * @param userToken {string}
      *
-     * @returns {Promise<boolean>}
+     * @returns {Promise<boolean|undefined>}
      */
-    async updateOneWorkspaceFromId(id, name, description) {
-        if (name === "" && description === "") {
-            return false;
+    async updateOneWorkspaceFromId(id, name, description, userToken) {
+        const user = await UserController.findOneUserFromToken(userToken);
+        if (user) {
+            if (name === "" && description === "") {
+                return false;
+            }
+            const workspace = await this.findOneWorkspaceFromId(id);
+            if (!workspace || workspace.user_id !== user.id) {
+                return false;
+            }
+            const values = {};
+            if (name !== workspace.name) {
+                values.name = name;
+            }
+            if (description !== workspace.description) {
+                values.description = description;
+            }
+            return await WorkspaceService.update(values, {id: id});
         }
-        const workspace = await this.findOneWorkspaceFromId(id);
-        if (!workspace) {
-            return false;
-        }
-        const values = {};
-        if (name !== workspace.name) {
-            values.name = name;
-        }
-        if (description !== workspace.description) {
-            values.description = description;
-        }
-        return await WorkspaceService.update(values, {id: id});
     }
 }
